dev-requirements.txt
pytest
pytest-django
black
isort
flake8
